cram windowing logic into new module
